# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api version 1",
        "description": "Laravel project API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "news list",
                "operationId": "31dcc9f6a1eb0da02d06f2b20e43d4e3",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "news limit"
                    },
                    {
                        "name": "user_email",
                        "in": "query",
                        "description": "емайл криейтера новости"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "параметр сортировки"
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "параметр сортировки"
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "дата от (1990-05-15)"
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "дата до (1990-05-15)"
                    },
                    {
                        "name": "tags[]",
                        "in": "query",
                        "description": "Список slug тегов. Пример: ?tags[]=impedit&tags[]=non",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "impedit"
                            }
                        }
                    },
                    {
                        "name": "tags_flag",
                        "in": "query",
                        "description": "флаг: true - у новости должны быть только переданные теги, false - один из тегов",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "enum": [
                                true,
                                false
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Добавить новость",
                "operationId": "13db4276bce4f49ebf340c0cafabccd8",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "title": {
                                        "description": "Название новости",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "content": {
                                        "description": "Текст",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "detail_picture": {
                                        "description": "Детальная картинка",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "preview_picture": {
                                        "description": "Картинка превью",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "images[]": {
                                        "description": "Массив изображений",
                                        "type": "array",
                                        "items": {
                                            "type": "file",
                                            "format": "binary"
                                        }
                                    },
                                    "tags": {
                                        "description": "теги",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "fail"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/v1/news/{slug}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "news by slug",
                "operationId": "d8d63491d4280161ab375527a74ccaa1",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "new slug"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Обновить указанные поля новости",
                "operationId": "c5d594c5d6122df7a896df063d8f2a72",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Символьный код новости"
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления (только изменяемые поля)",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Название новости",
                                        "type": "string",
                                        "example": "Новое название"
                                    },
                                    "content": {
                                        "description": "Текст новости",
                                        "type": "string",
                                        "example": "Обновлённый текст"
                                    },
                                    "detail_picture": {
                                        "description": "Детальная картинка",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "preview_picture": {
                                        "description": "Картинка превью",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "images[]": {
                                        "description": "Дополнительные изображения",
                                        "type": "array",
                                        "items": {
                                            "type": "file",
                                            "format": "binary"
                                        }
                                    },
                                    "tags": {
                                        "description": "теги",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "updated new"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Новость не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Удалить новость",
                "operationId": "3ce9bcaa45bfdcd40bbcbcd5a2507f78",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Символьный код новости"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}